requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
nbformat==5.9.2
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
pydantic==2.5.0
jinja2==3.1.2
